allan/exemplo_1_sonnet/requirements.txt
crewai>=0.28.8
langchain>=0.1.0
langchain-openai>=0.1.0
python-dotenv>=1.0.0
requests>=2.31.0
